# docs/swagger.yaml
basePath: /
definitions:
  models.BookingCreate:
    properties:
      end_time:
        description: YYYY-MM-DD-hh:mm:ss format
        type: string
      room_id:
        type: string
      start_time:
        description: YYYY-MM-DD-hh:mm:ss format
        type: string
    required:
    - end_time
    - room_id
    - start_time
    type: object
  models.BookingDTO:
    properties:
      end_time:
        type: string
      id:
        type: integer
      room:
        $ref: '#/definitions/models.RoomDTO'
      room_id:
        type: string
      start_time:
        type: string
    type: object
  models.RoomDTO:
    properties:
      id:
        type: string
      room_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for test_task_kami
  title: test_task_kami
  version: "1.0"
paths:
  /api/v1/bookings:
    post:
      consumes:
      - application/json
      description: Create a reservation for a room
      parameters:
      - description: BookingCreate - RoomID is required. StartTime and EndTime must
          be in the format YYYY-MM-DD-hh:mm:ss and are required.
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/models.BookingCreate'
      produces:
      - application/json
      responses:
        "201":
          description: id
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: invalid input
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create a new reservation
      tags:
      - Bookings
  /api/v1/bookings/{roomID}:
    get:
      description: Retrieve all reservations for a specific room
      parameters:
      - description: Room ID
        in: path
        name: roomID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BookingDTO'
            type: array
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get reservations by room ID
      tags:
      - Bookings
swagger: "2.0"